+++ basic_cms/__init__.py
@@ -1,9 +1,9 @@
 # -*- coding: utf-8 -*-
 """Django Basic CMS module."""
-VERSION = (0, 2, 4, 1)
+VERSION = (0, 3, 0)
 __version__ = '.'.join(map(str, VERSION))
-__author__ = "YD Technology"
-__contact__ = "team@ydtechnology.com"
+__author__ = "ArabellaTech"
+__contact__ = "geeks@arabel.la"
 __homepage__ = "https://github.com/YD-Technology/django-basic-cms"
 __docformat__ = "restructuredtext"
 __doc__ = 'A tree based Django Basic CMS application'