+++ basic_cms/managers.py
@@ -8,7 +8,10 @@ from django.db import models, connection
 from django.db.models import Q
 from django.core.cache import cache
 from django.core.exceptions import ObjectDoesNotExist
-from django.contrib.auth.models import SiteProfileNotAvailable
+try:
+    from django.contrib.auth.models import SiteProfileNotAvailable
+except ImportError:
+    SiteProfileNotAvailable = Exception
 from django.db.models import Avg, Max, Min, Count
 from django.contrib.sites.models import Site
 from django.conf import settings as global_settings
@@ -16,9 +19,15 @@ from django.utils.translation import ugettext_lazy as _
 
 import django
 
-if django.VERSION >= (1, 5):
+if django.VERSION >= (1, 5) and django.VERSION < (1, 7):
     from django.contrib.auth import get_user_model
     User = get_user_model()
+elif django.VERSION >= (1, 7):
+    # try:
+    #     from django.contrib.auth import get_user_model
+    #     User = settings.AUTH_USER_MODEL
+    # except AttributeError, ImportError:
+        from django.contrib.auth.models import User
 else:
     from django.contrib.auth.models import User
 
@@ -133,7 +142,6 @@ class PageManager(models.Manager):
                 return root_pages[0]
             else:
                 return None
-
         slug = get_slug(complete_path)
         from basic_cms.models import Content
         page_ids = Content.objects.get_page_ids_by_slug(slug)
@@ -142,8 +150,8 @@ class PageManager(models.Manager):
             pages_list = pages_list.exclude(status=self.model.DRAFT)
         if len(pages_list) == 1:
             if(settings.PAGE_USE_STRICT_URL and
-                pages_list[0].get_complete_slug(lang) != complete_path):
-                    return None
+                    pages_list[0].get_complete_slug(lang) != complete_path):
+                return None
             return pages_list[0]
         # if more than one page is matching the slug,
         # we need to use the full URL