+++ basic_cms/test_runner.py
@@ -11,7 +11,10 @@ current_dirname = os.path.dirname(__file__)
 sys.path.insert(0, current_dirname)
 sys.path.insert(0, os.path.join(current_dirname, '..'))
 
-from django.test.simple import DjangoTestSuiteRunner
+from django import setup
+setup()
+
+from django.test.runner import DiscoverRunner
 from django.db.models import get_app, get_apps
 import fnmatch
 
@@ -63,7 +66,7 @@ def get_all_coverage_modules(app_module, exclude_patterns=[]):
     return mod_list
 
 
-class PageTestSuiteRunner(DjangoTestSuiteRunner):
+class PageTestSuiteRunner(DiscoverRunner):
 
     def run_tests(self, test_labels=('basic_cms',), extra_tests=None):
 
@@ -73,7 +76,7 @@ class PageTestSuiteRunner(DjangoTestSuiteRunner):
             cov.use_cache(0)
             cov.start()
 
-        results = DjangoTestSuiteRunner.run_tests(self, test_labels, extra_tests)
+        results = DiscoverRunner.run_tests(self, test_labels, extra_tests)
 
         if coverage:
             cov.stop()
@@ -83,8 +86,8 @@ class PageTestSuiteRunner(DjangoTestSuiteRunner):
 
         sys.exit(results)
 
-def build_suite():
 
+def build_suite():
     runner = PageTestSuiteRunner()
     runner.setup_test_environment()
     runner.setup_databases()