+++ basic_cms/testproj/documents/models.py
@@ -3,6 +3,7 @@ from django.forms import ModelForm
 from basic_cms.models import Page
 from django.utils.translation import ugettext_lazy as _
 
+
 class Document(models.Model):
     "A dummy model used to illsutrate the use of linked models in django-page-cms"
 
@@ -12,7 +13,8 @@ class Document(models.Model):
     # the foreign key _must_ be called page
     page = models.ForeignKey(Page)
 
+
 class DocumentForm(ModelForm):
     class Meta:
         model = Document
-
+        exclude = []