+++ basic_cms/testproj/documents/views.py
@@ -4,7 +4,7 @@ from basic_cms.testproj.documents.models import Document
 
 
 def document_view(request, *args, **kwargs):
-    context = RequestContext(request, kwargs)
+    context = kwargs
     if kwargs.get('current_page', False):
         documents = Document.objects.filter(page=kwargs['current_page'])
         context['documents'] = documents
@@ -12,4 +12,6 @@ def document_view(request, *args, **kwargs):
         document = Document.objects.get(pk=int(kwargs['document_id']))
         context['document'] = document
     context['in_document_view'] = True
-    return render_to_response('pages/examples/index.html', context, context_instance=RequestContext(request))
+    rcontext = RequestContext(request, context)
+
+    return render_to_response('pages/examples/index.html', rcontext)