+++ basic_cms/testproj/test_settings.py
@@ -24,6 +24,9 @@ DATABASES = {
     }
 }
 
+# We still want to be ale to test with 1.1.X
+DATABASE_ENGINE = 'sqlite3'
+DATABASE_NAME = 'test.db'
 
 # Local time zone for this installation. Choices can be found here:
 # http://en.wikipedia.org/wiki/List_of_tz_zones_by_name
@@ -61,9 +64,23 @@ TEMPLATE_CONTEXT_PROCESSORS = (
     "django.core.context_processors.i18n",
     "django.core.context_processors.debug",
     "django.core.context_processors.request",
+    "django.contrib.messages.context_processors.messages",
     "django.core.context_processors.media",
     "basic_cms.context_processors.media",
 )
+
+# for django.1.8.x
+import django
+if django.VERSION >= (1, 8):
+    TEMPLATE_CONTEXT_PROCESSORS = (
+        "django.contrib.auth.context_processors.auth",
+        "django.template.context_processors.i18n",
+        "django.template.context_processors.debug",
+        "django.template.context_processors.request",
+        "django.contrib.messages.context_processors.messages",
+        "django.template.context_processors.media",
+        "basic_cms.context_processors.media",
+    )
 
 INTERNAL_IPS = ('127.0.0.1',)
 
@@ -72,7 +89,7 @@ MIDDLEWARE_CLASSES = (
     'django.middleware.locale.LocaleMiddleware',
     'django.middleware.common.CommonMiddleware',
     'django.contrib.auth.middleware.AuthenticationMiddleware',
-    'django.middleware.doc.XViewMiddleware',
+    # 'django.middleware.doc.XViewMiddleware',
     'django.contrib.messages.middleware.MessageMiddleware'
 )
 
@@ -85,6 +102,22 @@ TEMPLATE_DIRS = (
     os.path.join(PROJECT_DIR, 'templates'),
 )
 
+TEMPLATES = [
+    {
+        'BACKEND': 'django.template.backends.django.DjangoTemplates',
+        'DIRS': TEMPLATE_DIRS,
+        'APP_DIRS': True,
+        'OPTIONS': {
+            'context_processors': TEMPLATE_CONTEXT_PROCESSORS
+        },
+    },
+    # {
+    #     'BACKEND': 'django.template.backends.jinja2.Jinja2',
+    #     'DIRS': [
+    #         '/home/html/jinja2',
+    #     ],
+    # },
+]
 CACHE_BACKEND = "locmem:///?timeout=300&max_entries=6000"
 
 INSTALLED_APPS = (
@@ -95,15 +128,12 @@ INSTALLED_APPS = (
     'django.contrib.sites',
     'django.contrib.sitemaps',
     TEST_PROJ + '.documents',
+    'mptt',
     'taggit',
     'basic_cms',
-    'mptt',
     #'tinymce',
     # disabled to make "setup.py test" to work properly
     #'south',
-
-    # these 2 package don't create any dependecies
-    'authority',
     # haystack change coverage score report by importing modules
     #'haystack',
 )