+++ basic_cms/tests/__init__.py
@@ -3,6 +3,7 @@ import unittest
 from basic_cms.tests.test_functionnal import FunctionnalTestCase
 from basic_cms.tests.test_unit import UnitTestCase
 from basic_cms.tests.test_regression import RegressionTestCase
+from basic_cms.tests.test_api import CMSPagesApiTests
 
 
 def suite():
@@ -10,6 +11,7 @@ def suite():
     from basic_cms import settings
     if not settings.PAGE_ENABLE_TESTS:
         return suite
+    suite.addTest(unittest.makeSuite(CMSPagesApiTests))
     suite.addTest(unittest.makeSuite(UnitTestCase))
     suite.addTest(unittest.makeSuite(RegressionTestCase))
     # being the slower test I run it at the end