+++ basic_cms/urlconf_registry.py
@@ -1,5 +1,5 @@
 """Django page CMS urlconf registry."""
-from django.utils.translation import ugettext as _
+from django.utils.translation import ugettext_lazy as _
 
 
 class UrlconfAlreadyRegistered(Exception):