+++ example/settings.py
@@ -24,6 +24,10 @@ DATABASES = {
     }
 }
 
+# We still want to be ale to test with 1.1.X
+DATABASE_ENGINE = 'sqlite3'
+DATABASE_NAME = 'cms.db'
+
 # Local time zone for this installation. Choices can be found here:
 # http://en.wikipedia.org/wiki/List_of_tz_zones_by_name
 # although not all choices may be avilable on all operating systems.
@@ -94,15 +98,13 @@ INSTALLED_APPS = (
     'django.contrib.sites',
     'django.contrib.staticfiles',
     'pages.testproj.documents',
-    #'taggit',
+    # 'taggit',
     'pages',
     'mptt',
-    #'tinymce',
+    # 'tinymce',
     # disabled to make "setup.py test" to work properly
-    #'south',
+    # 'south',
 
-    # these 2 package don't create any dependecies
-    'authority',
     # haystack change coverage score report by importing modules
     'haystack',
 )
@@ -114,10 +116,10 @@ PAGE_CONNECTED_MODELS = [{
     'model': 'pages.testproj.documents.models.Document',
     'form': 'pages.testproj.documents.models.DocumentForm',
     'options': {
-            'extra': 3,
-            'max_num': 10,
-        },
-},]
+        'extra': 3,
+        'max_num': 10,
+    },
+}]
 
 # Default language code for this installation. All choices can be found here:
 # http://www.i18nguy.com/unicode/language-identifiers.html
@@ -141,6 +143,7 @@ languages = list(PAGE_LANGUAGES)
 languages.append(('fr-fr', gettext_noop('French')))
 languages.append(('fr-be', gettext_noop('Belgium french')))
 LANGUAGES = languages
+
 
 # This enable you to map a language(s) to another one, these languages should
 # be in the LANGUAGES config
@@ -190,4 +193,3 @@ try:
     from local_settings import *
 except ImportError:
     pass
-