+++ example/urls.py
@@ -12,7 +12,6 @@ admin.autodiscover()
 authority.autodiscover()
 
 urlpatterns = patterns('',
-    (r'^authority/', include('authority.urls')),
     (r'^i18n/', include('django.conf.urls.i18n')),
     (r'^admin/', include(admin.site.urls)),